refactor(runtime): isolate reconcile rollback authority
